requirements.txt
# --- AI & LangChain Ecosystem ---
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.34
langgraph>=0.0.38
openai>=1.20.0

# --- Vector Database ---
chromadb>=0.5.0

# --- Backend API (FastAPI) ---
fastapi>=0.110.1
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# --- Frontend  ---
streamlit>=1.33.0

# --- Data Processing & Utilities ---
pandas>=2.2.2
requests>=2.31.0
python-dotenv>=1.0.1


openpyxl